% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{cm_hr_model}
\alias{cm_hr_model}
\title{Combine a population with hazard ratios}
\usage{
cm_hr_model(population, hazard_ratios)
}
\arguments{
\item{population}{A \code{\link[=cm_population]{cm_population()}} (or \code{\link[=cm_model]{cm_model()}}, whose impacts are
ignored).}

\item{hazard_ratios}{A \code{\link[=cm_hazard_ratios]{cm_hazard_ratios()}} object with one value per
disease.}
}
\value{
A \code{cm_hr_model} object.
}
\description{
Combine a population with hazard ratios
}
