% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{joint_to_or}
\alias{joint_to_or}
\title{Odds ratio implied by a joint probability}
\usage{
joint_to_or(p11, p1, p2)
}
\arguments{
\item{p11}{Joint probability \verb{P(d1 and d2)}.}

\item{p1, p2}{Marginal probabilities.}
}
\value{
The odds ratio.
}
\description{
Odds ratio implied by a joint probability
}
