% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{set_three_way}
\alias{set_three_way}
\title{Set or replace one three-way association term}
\usage{
set_three_way(model, disease1, disease2, disease3, ratio)
}
\arguments{
\item{model}{A \code{\link[=cm_population]{cm_population()}} or \code{\link[=cm_model]{cm_model()}}.}

\item{disease1, disease2, disease3}{Disease ids.}

\item{ratio}{Ratio of conditional odds ratios (see \code{\link[=cm_three_way]{cm_three_way()}}).}
}
\value{
The modified object.
}
\description{
Set or replace one three-way association term
}
